LWDEV-4373: use float in SQL reports

+++ src/MarginTrading.Brokers/MarginTrading.AccountHistoryBroker/Repositories/Models/AccountTransactionsReport.cs
@@ -5,14 +5,14 @@ namespace MarginTrading.AccountHistoryBroker.Repositories.Models
     public class AccountTransactionsReport : IAccountTransactionsReport
     {
         public string AccountId { get; set; }
-        public decimal Amount { get; set; }
-        public decimal Balance { get; set; }
+        public double Amount { get; set; }
+        public double Balance { get; set; }
         public string ClientId { get; set; }
         public string Comment { get; set; }
         public DateTime Date { get; set; }
         public string Id { get; set; }
         public string PositionId { get; set; }
         public string Type { get; set; }
-        public decimal WithdrawTransferLimit { get; set; }
+        public double WithdrawTransferLimit { get; set; }
     }
 }
