+++ src/MarginTrading.Brokers/MarginTrading.AccountHistoryBroker/Repositories/Models/IAccountTransactionsReport.cs
@@ -5,14 +5,14 @@ namespace MarginTrading.AccountHistoryBroker.Repositories.Models
     public interface IAccountTransactionsReport
     {
         string AccountId { get; }
-        decimal Amount { get; }
-        decimal Balance { get; }
+        double Amount { get; }
+        double Balance { get; }
         string ClientId { get; }
         string Comment { get; }
         DateTime Date { get; }
         string Id { get; }
         string PositionId { get; }
         string Type { get; }
-        decimal WithdrawTransferLimit { get; }
+        double WithdrawTransferLimit { get; }
     }
 }
